+++ prefect_kubernetes/deployments.py
@@ -43,10 +43,10 @@ async def create_namespaced_deployment(
             )
         ```
     """
-    with kubernetes_credentials.get_client("core") as core_v1_client:
+    with kubernetes_credentials.get_client("apps") as apps_v1_client:
 
         return await run_sync_in_worker_thread(
-            core_v1_client.create_namespaced_deployment,
+            apps_v1_client.create_namespaced_deployment,
             namespace=namespace,
             body=new_deployment,
             **kube_kwargs,
@@ -91,10 +91,10 @@ async def delete_namespaced_deployment(
             )
         ```
     """
-    with kubernetes_credentials.get_client("core") as core_v1_client:
+    with kubernetes_credentials.get_client("apps") as apps_v1_client:
 
         return await run_sync_in_worker_thread(
-            core_v1_client.delete_namespaced_deployment,
+            apps_v1_client.delete_namespaced_deployment,
             deployment_name,
             body=delete_options,
             namespace=namespace,
@@ -133,10 +133,10 @@ async def list_namespaced_deployment(
             )
         ```
     """
-    with kubernetes_credentials.get_client("core") as core_v1_client:
+    with kubernetes_credentials.get_client("apps") as apps_v1_client:
 
         return await run_sync_in_worker_thread(
-            core_v1_client.list_namespaced_deployment,
+            apps_v1_client.list_namespaced_deployment,
             namespace=namespace,
             **kube_kwargs,
         )
@@ -180,10 +180,10 @@ async def patch_namespaced_deployment(
             )
         ```
     """
-    with kubernetes_credentials.get_client("core") as core_v1_client:
+    with kubernetes_credentials.get_client("apps") as apps_v1_client:
 
         return await run_sync_in_worker_thread(
-            core_v1_client.patch_namespaced_deployment,
+            apps_v1_client.patch_namespaced_deployment,
             name=deployment_name,
             namespace=namespace,
             body=deployment_updates,
@@ -224,10 +224,10 @@ async def read_namespaced_deployment(
             )
         ```
     """
-    with kubernetes_credentials.get_client("core") as core_v1_client:
+    with kubernetes_credentials.get_client("apps") as apps_v1_client:
 
         return await run_sync_in_worker_thread(
-            core_v1_client.read_namespaced_deployment,
+            apps_v1_client.read_namespaced_deployment,
             name=deployment_name,
             namespace=namespace,
             **kube_kwargs,
@@ -272,10 +272,10 @@ async def replace_namespaced_deployment(
             )
         ```
     """
-    with kubernetes_credentials.get_client("core") as core_v1_client:
+    with kubernetes_credentials.get_client("apps") as apps_v1_client:
 
         return await run_sync_in_worker_thread(
-            core_v1_client.replace_namespaced_deployment,
+            apps_v1_client.replace_namespaced_deployment,
             body=new_deployment,
             name=deployment_name,
             namespace=namespace,